requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.7.0
matplotlib>=3.4.0
PyWavelets>=1.1.0
scikit-learn>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
tqdm>=4.62.0
pytest>=7.0.0
pytest-cov>=3.0.0
